Move resources dir ; change project root
