+++ gazupublisher/resources/icons/app.icns
