+++ gazupublisher/resources/icons/app.ico
