+++ gazupublisher/resources/icons/refresh.png
