+++ gazupublisher/resources/resources.qrc
