+++ gazupublisher/resources/style.qss
