+++ gazupublisher/resources/translations/README.rst
