+++ gazupublisher/resources/views/TaskPanel.ui
