+++ gazupublisher/views/TasksTabItem.py
@@ -2,11 +2,11 @@ import Qt.QtWidgets as QtWidgets
 import Qt.QtGui as QtGui
 import Qt.QtCore as QtCore
 
-from gazupublisher.gazupublisher.utils.other import (
+from gazupublisher.utils.other import (
     combine_colors,
     from_min_to_day,
 )
-from gazupublisher.gazupublisher.utils.date import format_table_date
+from gazupublisher.utils.date import format_table_date
 
 
 class TasksTabItem(QtWidgets.QTableWidgetItem):
