+++ setup.py
@@ -3,7 +3,6 @@
 import re
 import os
 from subprocess import check_call
-
 from setuptools import setup, find_packages, Command
 from setuptools.command.sdist import sdist
 
@@ -30,36 +29,45 @@ with open("gazupublisher/__init__.py") as f:
     _version = re.search(r"__version__\s+=\s+\'(.*)\'", f.read()).group(1)
 
 
-if has_build_ui:
-
-    class build_res(build_ui):
-        """Build UI, resources and translations."""
-
-        def run(self):
-            # build translations
-            check_call(["pylupdate5", "app.pro"])
-
-            lrelease = os.environ.get("LRELEASE_BIN")
-            if not lrelease:
-                lrelease = "lrelease"
-
-            check_call([lrelease, "app.pro"])
-
-            # build UI & resources
-            build_ui.run(self)
-
-    cmdclass["build_res"] = build_res
-
-
-class custom_sdist(sdist):
-    """Custom sdist command."""
-
-    def run(self):
-        self.run_command("build_res")
-        sdist.run(self)
-
-
-cmdclass["sdist"] = custom_sdist
+# if has_build_ui:
+#
+#     class build_res(build_ui):
+#         """Build UI, resources and translations."""
+#
+#         def run(self):
+#             # build translations
+#             check_call(["pylupdate5", "app.pro"])
+#
+#             lrelease = os.environ.get("LRELEASE_BIN")
+#             if not lrelease:
+#                 lrelease = "lrelease"
+#
+#             check_call([lrelease, "app.pro"])
+#
+#             # build UI & resources
+#             build_ui.run(self)
+#
+#     cmdclass["build_res"] = build_res
+#
+#
+# class custom_build_py(build_py):
+#     def run(self):
+#         self.run_command('build_ui')
+#         build_py.run(self)
+#
+#
+# cmdclass["build_py"] = custom_build_py
+#
+#
+# class custom_sdist(sdist):
+#     """Custom sdist command."""
+#
+#     def run(self):
+#         self.run_command("build_res")
+#         sdist.run(self)
+#
+#
+# cmdclass["sdist"] = custom_sdist
 
 
 class bdist_app(Command):
@@ -87,12 +95,10 @@ install_requirements = [
     "qt.py@git+https://github.com/mottosso/Qt.py.git#egg=qt.py",
 ]
 
-project_root = os.path.normpath(os.path.dirname(__file__))
-
 setup(
     name="gazupublisher",
     version=_version,
-    packages=find_packages(".."),
+    packages=find_packages(),
     description="Application to publish previews to Kitsu from desktop environments",
     author="CGWire",
     author_email="dev@cg-wire.com",
@@ -103,4 +109,6 @@ setup(
     },
     cmdclass=cmdclass,
     install_requires=install_requirements,
+    package_data={'resources': ['*']},
+    include_package_data=True,
 )
